Move `createGroup` into `groups_Create`
